Extract common methods from functional tests
