+++ tests/commonTemplates_test.go
@@ -1,51 +1,51 @@
 package tests
 
 import (
+	"reflect"
+
 	. "github.com/onsi/ginkgo"
 	"github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 	templatev1 "github.com/openshift/api/template/v1"
 	core "k8s.io/api/core/v1"
 	rbac "k8s.io/api/rbac/v1"
-	commonTemplates "kubevirt.io/ssp-operator/internal/operands/common-templates"
-	"reflect"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"time"
-)
 
-var (
-	viewRole = &testResource{
-		Name:       commonTemplates.ViewRoleName,
-		Namsespace: commonTemplates.GoldenImagesNSname,
-		resource:   &rbac.Role{},
-	}
-	viewRoleBinding = &testResource{
-		Name:       commonTemplates.ViewRoleName,
-		Namsespace: commonTemplates.GoldenImagesNSname,
-		resource:   &rbac.RoleBinding{},
-	}
-	editClusterRole = &testResource{
-		Name:       commonTemplates.EditClusterRoleName,
-		resource:   &rbac.ClusterRole{},
-		Namsespace: "",
-	}
-	goldenImageNS = &testResource{
-		Name:       commonTemplates.GoldenImagesNSname,
-		resource:   &core.Namespace{},
-		Namsespace: "",
-	}
-	testTemplate = &testResource{
-		Name:       "centos6-server-large-v0.11.3",
-		Namsespace: commonTemplatesTestNS,
-		resource:   &templatev1.Template{},
-	}
+	commonTemplates "kubevirt.io/ssp-operator/internal/operands/common-templates"
 )
 
 var _ = Describe("Common templates", func() {
+	var (
+		viewRole = &testResource{
+			Name:       commonTemplates.ViewRoleName,
+			Namsespace: commonTemplates.GoldenImagesNSname,
+			resource:   &rbac.Role{},
+		}
+		viewRoleBinding = &testResource{
+			Name:       commonTemplates.ViewRoleName,
+			Namsespace: commonTemplates.GoldenImagesNSname,
+			resource:   &rbac.RoleBinding{},
+		}
+		editClusterRole = &testResource{
+			Name:       commonTemplates.EditClusterRoleName,
+			resource:   &rbac.ClusterRole{},
+			Namsespace: "",
+		}
+		goldenImageNS = &testResource{
+			Name:       commonTemplates.GoldenImagesNSname,
+			resource:   &core.Namespace{},
+			Namsespace: "",
+		}
+		testTemplate = &testResource{
+			Name:       "centos6-server-large-v0.11.3",
+			Namsespace: commonTemplatesTestNS,
+			resource:   &templatev1.Template{},
+		}
+	)
+
 	Context("resource creation", func() {
 		table.DescribeTable("created cluster resource", func(res *testResource) {
 			resource := res.NewResource()
-			err := apiClient.Get(ctx, client.ObjectKey{Name: res.Name}, resource)
+			err := apiClient.Get(ctx, res.GetKey(), resource)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(hasOwnerAnnotations(resource.GetAnnotations())).To(BeTrue())
 		},
@@ -54,9 +54,7 @@ var _ = Describe("Common templates", func() {
 		)
 
 		table.DescribeTable("created namespaced resource", func(res *testResource) {
-			err := apiClient.Get(ctx, client.ObjectKey{
-				Name: res.Name, Namespace: res.Namsespace,
-			}, res.NewResource())
+			err := apiClient.Get(ctx, res.GetKey(), res.NewResource())
 			Expect(err).ToNot(HaveOccurred())
 		},
 			table.Entry("[test_id:4777]view role", viewRole),
@@ -64,37 +62,13 @@ var _ = Describe("Common templates", func() {
 		)
 
 		It("[test_id:5086]Create common-template in custom NS", func() {
-			err := apiClient.Get(ctx, client.ObjectKey{
-				Name: testTemplate.Name, Namespace: commonTemplatesTestNS,
-			}, testTemplate.NewResource())
+			err := apiClient.Get(ctx, testTemplate.GetKey(), testTemplate.NewResource())
 			Expect(err).ToNot(HaveOccurred())
 		})
 	})
 
 	Context("resource change", func() {
-		table.DescribeTable("should restore modified resource", func(
-			res *testResource,
-			updateFunc interface{},
-			equalsFunc interface{},
-		) {
-			key := res.GetKey()
-			original := res.NewResource()
-			Expect(apiClient.Get(ctx, key, original)).ToNot(HaveOccurred())
-
-			changed := original.DeepCopyObject()
-			reflect.ValueOf(updateFunc).Call([]reflect.Value{reflect.ValueOf(changed)})
-			Expect(apiClient.Update(ctx, changed)).ToNot(HaveOccurred())
-
-			newRes := res.NewResource()
-			Eventually(func() bool {
-				Expect(apiClient.Get(ctx, key, newRes)).ToNot(HaveOccurred())
-				res := reflect.ValueOf(equalsFunc).Call([]reflect.Value{
-					reflect.ValueOf(original),
-					reflect.ValueOf(newRes),
-				})
-				return res[0].Interface().(bool)
-			}, timeout, time.Second).Should(BeTrue())
-		},
+		table.DescribeTable("should restore modified resource", expectRestoreAfterUpdate,
 			table.Entry("edit cluster role", editClusterRole,
 				func(role *rbac.ClusterRole) {
 					role.Rules[0].Verbs = []string{"watch"}
@@ -130,18 +104,7 @@ var _ = Describe("Common templates", func() {
 	})
 
 	Context("resource deletion", func() {
-		table.DescribeTable("recreate after delete", func(res *testResource) {
-			resource := res.NewResource()
-			resource.SetName(res.Name)
-			resource.SetNamespace(res.Namsespace)
-			Expect(apiClient.Delete(ctx, resource)).ToNot(HaveOccurred())
-
-			Eventually(func() error {
-				return apiClient.Get(ctx, client.ObjectKey{
-					Name: res.Name, Namespace: res.Namsespace,
-				}, resource)
-			}, timeout, time.Second).ShouldNot(HaveOccurred())
-		},
+		table.DescribeTable("recreate after delete", expectRecreateAfterDelete,
 			table.Entry("[test_id:4773]view role", viewRole),
 			table.Entry("[test_id:4842]view role binding", viewRoleBinding),
 			table.Entry("[test_id:5088]testTemplate in custom NS", testTemplate),
