+++ tests/tests_suite_test.go
@@ -2,14 +2,14 @@ package tests
 
 import (
 	"context"
-	secv1 "github.com/openshift/api/security/v1"
-	templatev1 "github.com/openshift/api/template/v1"
 	"testing"
 	"time"
 
 	promv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	secv1 "github.com/openshift/api/security/v1"
+	templatev1 "github.com/openshift/api/template/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
